Cast seconds to int in uptime calculation
Ensures the seconds value in the uptime command is always an integer by explicitly casting the result of floor() to int.

+++ Commands/Uptime.php
@@ -11,7 +11,7 @@
 		$diff -= $hours * 3600;
 		$minutes = floor($diff / 60) % 60;
 		$diff -= $minutes * 60;
-		$seconds = floor($diff % 60);
+		$seconds = (int) floor($diff % 60);
 		$message->reply("{$days} days, {$hours} hrs, {$minutes} mins, {$seconds} secs");
 		
 	}
